move to text.py
