+++ rdt/transformers/text.py
@@ -2,10 +2,70 @@
 import warnings
 
 import numpy as np
+import pandas as pd
 
 from rdt.errors import TransformerProcessingError
 from rdt.transformers.base import BaseTransformer
 from rdt.transformers.utils import strings_from_regex
+
+
+class IDGenerator(BaseTransformer):
+    """Generate an ID column.
+
+    This transformer generates an ID column based on a given prefix, starting value and suffix.
+
+    Args:
+            prefix (str):
+                Prefix of the generated IDs column.
+                Defaults to ``None``.
+            starting_value (int):
+                Starting value of the generated IDs column.
+                Defaults to ``0``.
+            suffix (str):
+                Suffix of the generated IDs column.
+                Defaults to ``None``.
+    """
+
+    IS_GENERATOR = True
+    INPUT_SDTYPE = 'text'
+
+    def __init__(self, prefix=None, starting_value=0, suffix=None):
+        super().__init__()
+        self.prefix = prefix
+        self.starting_value = starting_value
+        self.suffix = suffix
+        self._counter = 0
+        self.output_properties = {None: {'next_transformer': None}}
+
+    def reset_randomization(self):
+        """Reset the sampling _counter."""
+        self._counter = 0
+
+    def _fit(self, data):
+        pass
+
+    def _transform(self, _data):
+        """Drop the input column by returning ``None``."""
+        return None
+
+    def _reverse_transform(self, data):
+        """Generate new id column.
+
+        Args:
+            data (pd.Series or numpy.ndarray):
+                Data to transform.
+
+        Returns:
+            pd.Series
+        """
+        start = self.starting_value + self._counter
+        prefix_str = self.prefix if self.prefix is not None else ''
+        suffix_str = self.suffix if self.suffix is not None else ''
+
+        values = [f'{prefix_str}{start + idx}{suffix_str}' for idx in range(len(data))]
+        self._counter += len(data)
+
+        return pd.Series(values)
 
 
 class RegexGenerator(BaseTransformer):
